Extract module for entry point management.
